+++ src/use_locales.rs
@@ -39,7 +39,7 @@ use std::rc::Rc;
 /// ### Bring your own header
 ///
 /// In case you're neither using Axum, Actix nor Spin, or the default implementation is not to your liking,
-/// you can provide your own way of reading and writing the cookie header value using the option
+/// you can provide your own way of reading the language header value using the option
 /// [`crate::UseLocalesOptions::ssr_lang_header_getter`].
 pub fn use_locales() -> Signal<Vec<String>> {
     use_locales_with_options(UseLocalesOptions::default())